Update support for k8s 0.19.0
Here we've moved all the dns config to v1beta3 otherwise it won't work.
In order to work with the new version, we need to update kube2sky to the latest version.
We need to give some configurations to and it depends k8s new security features. 
We don't need to deal with them.
So, we bind the API server to all Ips. Then we connect to api server inside from the kube3sky 
via the docker's bridge ip. It's kind of a hack, but working well. 
Since this is only for dev purpose, we are good to go.